Added more comment

+++ image_to_vid.py
@@ -9,6 +9,6 @@ root =Tk()
 root.title("Image merger") #Set title
 root.geometry('1024x768+500+100') #set size of window and location it appears
 root.resizable(width=True, height=True)#Prevent user from changing size of window
-rClickbinder(root)#Enable copy paste
+rClickbinder(root)#Enable copy paste as long
 
 root.mainloop()
